Move settings file to top level
