+++ src/dev/map-layers/catalog.ts
@@ -5,7 +5,7 @@ import { Routes, Route, Link, useParams } from "react-router-dom";
 import styles from "../main.module.styl";
 import { useAPIResult, ErrorBoundary } from "@macrostrat/ui-components";
 import { ParentRouteButton } from "~/components/map-navbar";
-import { SETTINGS } from "~/pages/map/map-interface/settings";
+import { SETTINGS } from "~/settings";
 import { Spinner } from "@blueprintjs/core";
 import { BasicLayerInspectorPage } from ".";
