+++ src/dev/map-layers/index.ts
@@ -11,7 +11,7 @@ import {
 import { useStoredState, useDarkMode } from "@macrostrat/ui-components";
 import mapboxgl from "mapbox-gl";
 import { useCallback, useEffect, useState, useMemo } from "react";
-import { SETTINGS } from "~/pages/map/map-interface/settings";
+import { SETTINGS } from "~/settings";
 import { FloatingNavbar } from "@macrostrat/map-interface";
 import { useAppActions } from "../../pages/map/map-interface/app-state";
 import { LocationPanel } from "@macrostrat/map-interface";
