+++ src/pages/map/map-interface/app-state/handlers/fetch.ts
@@ -1,6 +1,6 @@
 import axios from "axios";
 import { joinURL } from "~/pages/map/map-interface/utils";
-import { SETTINGS } from "../../settings";
+import { SETTINGS } from "../../../../../settings";
 import { ColumnGeoJSONRecord } from "../reducers";
 import { UPDATE_FILTERED_COLUMNS } from "../reducers/filtered-columns";
