+++ src/pages/map/map-interface/app-state/nav-hooks.ts
@@ -2,7 +2,7 @@ import { useLocation, useNavigate } from "react-router";
 import { useAppState } from "./hooks";
 import classNames from "classnames";
 import { MenuPage } from "./reducers";
-import { routerBasename, mapPagePrefix } from "../settings";
+import { routerBasename, mapPagePrefix } from "../../../../settings";
 
 export function isDetailPanelRouteInternal(pathname: string) {
   /* Check if we're in a detail panel route from within the app. */
