+++ src/pages/map/map-interface/components/docs.ts
@@ -3,7 +3,7 @@ import { Alignment } from "@blueprintjs/core";
 import styles from "./docs.module.styl";
 import { useInView } from "react-intersection-observer";
 import { HashLink } from "react-router-hash-link";
-import { routerBasename } from "../settings";
+import { routerBasename } from "../../../../settings";
 import classNames from "classnames";
 import { joinURL } from "~/pages/map/map-interface/utils";
