+++ src/pages/map/map-interface/components/elevation-chart/index.ts
@@ -14,7 +14,7 @@ import { useAPIResult } from "@macrostrat/ui-components";
 import { LocationFocusButton } from "@macrostrat/mapbox-react";
 
 import styles from "./main.module.styl";
-import { SETTINGS } from "~/pages/map/map-interface/settings";
+import { SETTINGS } from "~/settings";
 const h = hyper.styled(styles);
 
 function drawElevationChart(
