+++ src/pages/map/map-interface/map-page/menu.ts
@@ -35,7 +35,7 @@ import { isDetailPanelRouteInternal } from "../app-state/nav-hooks";
 import { SettingsPanel, ExperimentsPanel, ThemeButton } from "./settings-panel";
 import { useState, useEffect } from "react";
 import { LinkButton, LayerButton, ListButton } from "../components/buttons";
-import { routerBasename, mapPagePrefix } from "../settings";
+import { routerBasename, mapPagePrefix } from "../../../../settings";
 import { Card } from "@blueprintjs/core";
 
 function ChangelogPanel() {
